+++ src/main/java/org/example/multithreading/basics/Consumer.java
@@ -1,4 +1,4 @@
-package org.example.multithreading;
+package org.example.multithreading.basics;
 
 import java.util.concurrent.BlockingQueue;
