+++ src/main/java/org/example/multithreading/basics/ConsumerThread.java
@@ -1,4 +1,4 @@
-package org.example.multithreading;
+package org.example.multithreading.basics;
 
 public class ConsumerThread implements  Runnable{
